app/backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# GCP clients
google-cloud-storage==2.14.0
google-cloud-bigquery==3.14.1
google-cloud-pubsub==2.19.0
google-cloud-secret-manager==2.17.0
google-cloud-documentai==2.24.0
google-cloud-aiplatform==1.40.0
google-cloud-tasks==2.15.0
google-cloud-logging==3.9.0

# Excel/PDF processing
openpyxl==3.1.2
XlsxWriter==3.1.9
pandas==2.1.4
numpy==1.26.3
pypdf==4.0.0
pymupdf==1.23.8
pdfminer.six==20231228

# PDF generation
weasyprint==60.2
jinja2==3.1.3

# HTTP client
httpx==0.26.0
aiohttp==3.9.1

# Data validation
email-validator==2.1.0

# Market data
yfinance==0.2.36

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Testing (dev)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.1.0

# Code quality (dev)
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Observability
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-gcp-trace==1.6.0
